Moved and improved leveldb/plyvel tests
Fixed import error in Plyvel tests

Improved LevelDB tests
